chore(solace): extract another function
