+++ src/apps/solace/ethereum/solace.tvl-fetcher.ts
@@ -4,15 +4,11 @@ import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
 import { TvlFetcher } from '~stats/tvl/tvl-fetcher.interface';
 import { Network } from '~types/network.interface';
-import { Token } from '~position/position.interface';
-import { WithMetaType } from '~position/display.interface';
 
 import { SolaceContractFactory } from '../contracts';
 import { SOLACE_DEFINITION } from '../solace.definition';
 
-import { ethers } from 'ethers';
-const formatUnits = ethers.utils.formatUnits;
-import { bnToFloat, range } from '~apps/solace/utils';
+import { bnToFloat, findToken, range } from '~apps/solace/utils';
 
 const appId = SOLACE_DEFINITION.id;
 const network = Network.ETHEREUM_MAINNET;
@@ -20,10 +16,6 @@ const network = Network.ETHEREUM_MAINNET;
 // TODO: there is more to TVL than amount in the UWP
 // since its the bulk we call it sufficient for now
 const UWP_ADDRESS    = "0x5efC0d9ee3223229Ce3b53e441016efC5BA83435";
-const SOLACE_ADDRESS = "0x501ace9c35e60f03a2af4d484f49f9b1efde9f40";
-const SCP_ADDRESS    = "0x501acee83a6f269b77c167c6701843d454e2efa0";
-const SLP_ADDRESS    = "0x9c051f8a6648a51ef324d30c235da74d060153ac";
-const USDC_ADDRESS   = "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48";
 const ZERO_ADDRESS   = "0x0000000000000000000000000000000000000000";
 const TOKENS = [
   {
@@ -102,55 +94,10 @@ export class EthereumSolaceTvlFetcher implements TvlFetcher {
     }));
     let usd = 0;
     await Promise.all(indices.map(async (i:number) => {
-      const zapperToken = await this.getToken(baseTokens, TOKENS[i].address);
+      const zapperToken = await findToken(baseTokens, TOKENS[i].address, this.solaceContractFactory, network);
       if(!zapperToken) return;
       usd += balances[i] * zapperToken.price;
     }));
     return usd;
-  }
-
-  async getToken(baseTokens:WithMetaType<Token>[], tokenAddress:string) {
-    const token = baseTokens.find((t:WithMetaType<Token>) => t.address === tokenAddress);
-    if(!!token) return token;
-    if(tokenAddress === SCP_ADDRESS) {
-      const scp = this.solaceContractFactory.scp({ address: SCP_ADDRESS, network });
-      const pps = await scp.pricePerShare();
-      const eth = baseTokens.find((t:WithMetaType<Token>) => t.address === ZERO_ADDRESS);
-      const ethPrice = eth?.price ?? 0.0;
-      const scpPrice = ethPrice * bnToFloat(18)(pps);
-      return {
-        "metaType": "supplied",
-        "type": "app-token",
-        "network": "ethereum",
-        "address": SCP_ADDRESS,
-        "decimals": 18,
-        "symbol": "SCP",
-        "price": scpPrice
-      }
-    } else if(tokenAddress === SLP_ADDRESS) {
-      const solace = baseTokens.find((t:WithMetaType<Token>) => t.address === SOLACE_ADDRESS);
-      const usdc = baseTokens.find((t:WithMetaType<Token>) => t.address === USDC_ADDRESS);
-      const solacePrice = solace?.price ?? 0.0;
-      const usdcPrice = usdc?.price ?? 0.0;
-      const solaceContract = this.solaceContractFactory.erc20({ address: SOLACE_ADDRESS, network });
-      const usdcContract = this.solaceContractFactory.erc20({ address: USDC_ADDRESS, network });
-      const slpContract = this.solaceContractFactory.erc20({ address: SLP_ADDRESS, network });
-      const [s, u, ts] = await Promise.all([
-        solaceContract.balanceOf(SLP_ADDRESS).then(bnToFloat(18)),
-        usdcContract.balanceOf(SLP_ADDRESS).then(bnToFloat(6)),
-        slpContract.totalSupply().then(bnToFloat(18)),
-      ])
-      const slpPrice = (s*solacePrice + u*usdcPrice) / ts;
-      return {
-        "metaType": "supplied",
-        "type": "app-token",
-        "network": "ethereum",
-        "address": SLP_ADDRESS,
-        "decimals": 18,
-        "symbol": "SLP",
-        "price": slpPrice
-      }
-    }
-    return undefined;
   }
 }
