+++ src/apps/solace/polygon/solace.tvl-fetcher.ts
@@ -4,13 +4,11 @@ import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
 import { TvlFetcher } from '~stats/tvl/tvl-fetcher.interface';
 import { Network } from '~types/network.interface';
-import { Token } from '~position/position.interface';
-import { WithMetaType } from '~position/display.interface';
 
 import { SolaceContractFactory } from '../contracts';
 import { SOLACE_DEFINITION } from '../solace.definition';
 
-import { bnToFloat, range } from '~apps/solace/utils';
+import { bnToFloat, findToken, range } from '~apps/solace/utils';
 
 const appId = SOLACE_DEFINITION.id;
 const network = Network.POLYGON_MAINNET;
@@ -92,16 +90,10 @@ export class PolygonSolaceTvlFetcher implements TvlFetcher {
     }));
     let usd = 0;
     await Promise.all(indices.map(async (i:number) => {
-      const zapperToken = await this.getToken(baseTokens, TOKENS[i].address);
+      const zapperToken = await findToken(baseTokens, TOKENS[i].address, this.solaceContractFactory, network);
       if(!zapperToken) return;
       usd += balances[i] * zapperToken.price;
     }));
     return usd;
-  }
-
-  async getToken(baseTokens:WithMetaType<Token>[], tokenAddress:string) {
-    const token = baseTokens.find((t:WithMetaType<Token>) => t.address === tokenAddress);
-    if(!!token) return token;
-    return undefined;
   }
 }
